# CLDC-2831 Update logs ordering